Add the channel_configuration support to control how the channel data is read (live or from log file) as well as if to reset the channel log file.
The Channel DTO model now caches the entire log file content if "expand" is requested.  Each read updates the internal cache then decimates the data before it's returned (so a smaller amount of data gets placed in the memcache).

Eventually I want to add the decimation as a configuration parameter to the request (and part of the channels_config object).

The hwloop prints the channels_config (if it's present) state.