Update the size for the log files.
There may be an issue with the memcache item size (default is 1MB, but I've modified the /etc/memcache.conf file to increase this to 10MB (the switch in the conf file that says -I 10m).  More testing is needed, but the memcache writes silently fail if the size of the item being written is too big (which may occur if channel log history is requested).

An alternative approach should probably be investigated once we start looking at logs for controls as well as sensors.