Updates the hardware loop and DTO channel model so as to split apart the data collection and DTO update process from the DTO publishing step.
Channel data is published (to memcache) using a format that puts each channel as a top-level object.  Then each DTO channel can be responsible for writing its own data to the cache and thus can be threaded so as not to block other channel updates during long publishing operations (e.g., when the channel is publishing its entire log history).  In addition a top-level list of active channels is also published on each cycle through the hardware loop.

Finally, when requested, a DTO channel will retrieve it's history from file for the first such request, but just update an in-memory copy with latest data points for subsequent requests to avoid file I/O bottleneck.